Fixed bugs for long enum and IEnumerable

+++ src/SwaggerWcf/Models/DefinitionProperty.cs
@@ -53,7 +53,7 @@ namespace SwaggerWcf.Models
 
         public bool UniqueItems { get; set; }
 
-        public List<int> Enum { get; set; }
+        public List<long> Enum { get; set; }
 
         public decimal MultipleOf { get; set; }
