+++ src/SwaggerWcf/Support/TypePropertiesProcessor.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
+using System.Diagnostics;
 using System.Reflection;
 using System.Runtime.Serialization;
 using SwaggerWcf.Attributes;
@@ -62,7 +63,7 @@ namespace SwaggerWcf.Support
 
         private static DefinitionProperty ProcessProperty(PropertyInfo propertyInfo, IList<string> hiddenTags,
                                                           Stack<Type> typesStack)
-        {
+        { 
             if (propertyInfo.GetCustomAttribute<SwaggerWcfHiddenAttribute>() != null
                 || propertyInfo.GetCustomAttributes<SwaggerWcfTagAttribute>()
                                .Select(t => t.TagName)
@@ -130,7 +131,7 @@ namespace SwaggerWcf.Support
 
             if ((prop.TypeFormat.Type == ParameterType.Integer && prop.TypeFormat.Format == "enum") || (prop.TypeFormat.Type == ParameterType.Array && prop.Items.TypeFormat.Format == "enum"))
             {
-                prop.Enum = new List<int>();
+                prop.Enum = new List<long>();
 
                 Type propType = propertyInfo.PropertyType;
 
@@ -144,7 +145,7 @@ namespace SwaggerWcf.Support
                     var enumMemberItem = Enum.Parse(propType, enumName, true);
                     string enumMemberDescription = DefinitionsBuilder.GetEnumDescription((Enum)enumMemberItem);
                     enumMemberDescription = (string.IsNullOrWhiteSpace(enumMemberDescription)) ? "" : $"({enumMemberDescription})";
-                    int enumMemberValue = DefinitionsBuilder.GetEnumMemberValue(propType, enumName);
+                    long enumMemberValue = DefinitionsBuilder.GetEnumMemberValue(propType, enumName);
                     if (prop.Description != null) prop.Enum.Add(enumMemberValue);
                     enumDescription += $"    {enumName}{System.Web.HttpUtility.HtmlEncode(" = ")}{enumMemberValue} {enumMemberDescription}\r\n";
                 }
